requirements.txt
spotipy>=2.23.0
requests>=2.31.0
apscheduler>=3.10.0
geopy>=2.4.0
pyyaml>=6.0
rich>=13.0.0
flask>=3.0.0
pytest>=7.0.0
pytest-mock>=3.0.0
